hal/requirements_macos.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements_macos.txt requirements.in
#
annotated-types==0.7.0
    # via pydantic
attrs==25.1.0
    # via
    #   -r requirements.in
    #   yasoo
autoflake==2.1.1
    # via -r requirements.in
autotyping==23.3.0
    # via -r requirements.in
azure-core==1.32.0
    # via
    #   azure-identity
    #   azure-storage-blob
    #   azure-storage-file-datalake
azure-identity==1.20.0
    # via mosaicml-streaming
azure-storage-blob==12.24.1
    # via
    #   azure-storage-file-datalake
    #   mosaicml-streaming
azure-storage-file-datalake==12.18.1
    # via mosaicml-streaming
bcrypt==4.3.0
    # via paramiko
black==23.9.1
    # via -r requirements.in
boto3==1.37.9
    # via mosaicml-streaming
botocore==1.37.9
    # via
    #   boto3
    #   s3transfer
brotli==1.1.0
    # via mosaicml-streaming
build==1.2.2.post1
    # via pip-tools
cachetools==5.5.2
    # via google-auth
catalogue==2.0.10
    # via mosaicml-streaming
certifi==2025.1.31
    # via
    #   oci
    #   requests
    #   sentry-sdk
cffi==1.17.1
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.4.1
    # via requests
circuitbreaker==2.0.0
    # via oci
click==8.1.8
    # via
    #   -r requirements.in
    #   black
    #   pip-tools
    #   wandb
cloudpickle==3.1.1
    # via tensordict
contourpy==1.3.1
    # via matplotlib
cramjam==2.9.1
    # via python-snappy
cryptography==44.0.2
    # via
    #   azure-identity
    #   azure-storage-blob
    #   msal
    #   oci
    #   paramiko
    #   pyjwt
    #   pyopenssl
cycler==0.12.1
    # via matplotlib
cython==0.29.37
    # via pyenet
docker-pycreds==0.4.0
    # via wandb
filelock==3.17.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fonttools==4.56.0
    # via matplotlib
fsspec==2025.3.0
    # via
    #   huggingface-hub
    #   torch
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via wandb
google-api-core==2.24.1
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.38.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.4.2
    # via google-cloud-storage
google-cloud-storage==2.10.0
    # via mosaicml-streaming
google-crc32c==1.6.0
    # via google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.69.1
    # via google-api-core
huggingface-hub==0.29.2
    # via
    #   tokenizers
    #   transformers
idna==3.10
    # via requests
iniconfig==2.0.0
    # via pytest
isodate==0.7.2
    # via
    #   azure-storage-blob
    #   azure-storage-file-datalake
isort==6.0.1
    # via -r requirements.in
jinja2==3.1.6
    # via torch
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
joblib==1.4.2
    # via scikit-learn
kiwisolver==1.4.8
    # via matplotlib
libcst==1.6.0
    # via
    #   -r requirements.in
    #   autotyping
loguru==0.7.3
    # via -r requirements.in
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.1
    # via
    #   -r requirements.in
    #   mosaicml-streaming
    #   seaborn
melee @ git+https://github.com/vladfi1/libmelee@b4a1a2585119bcedc398a512990dd9b95102eb74
    # via -r requirements.in
more-itertools==10.6.0
    # via
    #   -r requirements.in
    #   yasoo
mosaicml-streaming==0.11.0
    # via -r requirements.in
mpmath==1.3.0
    # via sympy
msal==1.31.1
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.2.0
    # via azure-identity
mypy-extensions==1.0.0
    # via
    #   -r requirements.in
    #   black
networkx==3.4.2
    # via torch
numpy==2.1.3
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
    #   melee
    #   mosaicml-streaming
    #   pandas
    #   pyarrow
    #   scikit-learn
    #   scipy
    #   seaborn
    #   tensordict
    #   torchvision
    #   transformers
nvidia-ml-py==12.570.86
    # via nvitop
nvitop==1.4.2
    # via -r requirements.in
oci==2.147.0
    # via mosaicml-streaming
orjson==3.10.15
    # via tensordict
packaging==24.2
    # via
    #   black
    #   build
    #   huggingface-hub
    #   matplotlib
    #   melee
    #   pytest
    #   transformers
pandas==2.2.3
    # via
    #   -r requirements.in
    #   seaborn
paramiko==3.5.1
    # via mosaicml-streaming
pathspec==0.12.1
    # via black
pillow==11.1.0
    # via
    #   matplotlib
    #   torchvision
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.3.6
    # via
    #   black
    #   wandb
pluggy==1.5.0
    # via pytest
portalocker==2.10.1
    # via msal-extensions
proto-plus==1.26.0
    # via google-api-core
protobuf==5.29.3
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
    #   wandb
psutil==7.0.0
    # via
    #   nvitop
    #   wandb
py-ubjson==0.16.1
    # via melee
pyarrow==16.0.0
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pycparser==2.22
    # via cffi
pydantic==2.10.6
    # via wandb
pydantic-core==2.27.2
    # via pydantic
pyenet @ git+https://github.com/piqueserver/pyenet
    # via melee
pyflakes==3.2.0
    # via autoflake
pyjwt[crypto]==2.10.1
    # via
    #   msal
    #   pyjwt
pynacl==1.5.0
    # via paramiko
pyopenssl==24.3.0
    # via oci
pyparsing==3.2.1
    # via matplotlib
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==8.3.5
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   matplotlib
    #   oci
    #   pandas
python-snappy==0.7.3
    # via mosaicml-streaming
pytz==2025.1
    # via
    #   oci
    #   pandas
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   libcst
    #   transformers
    #   wandb
regex==2024.11.6
    # via transformers
requests==2.32.3
    # via
    #   azure-core
    #   google-api-core
    #   google-cloud-storage
    #   huggingface-hub
    #   msal
    #   transformers
    #   wandb
rsa==4.9
    # via google-auth
s3transfer==0.11.4
    # via boto3
safetensors==0.5.3
    # via transformers
scikit-learn==1.6.1
    # via -r requirements.in
scipy==1.15.2
    # via scikit-learn
seaborn==0.13.2
    # via -r requirements.in
sentry-sdk==2.22.0
    # via wandb
setproctitle==1.3.5
    # via wandb
six==1.17.0
    # via
    #   azure-core
    #   docker-pycreds
    #   python-dateutil
smmap==5.0.2
    # via gitdb
sympy==1.13.1
    # via torch
tensordict==0.6.0
    # via -r requirements.in
threadpoolctl==3.5.0
    # via scikit-learn
tokenizers==0.21.0
    # via transformers
torch==2.6.0
    # via
    #   -r requirements.in
    #   mosaicml-streaming
    #   tensordict
    #   torchvision
torchvision==0.21.0
    # via
    #   -r requirements.in
    #   mosaicml-streaming
tqdm==4.67.1
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   mosaicml-streaming
    #   transformers
transformers==4.49.0
    # via mosaicml-streaming
typing-extensions==4.12.2
    # via
    #   azure-core
    #   azure-identity
    #   azure-storage-blob
    #   azure-storage-file-datalake
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   torch
    #   wandb
tzdata==2025.1
    # via pandas
urllib3==2.3.0
    # via
    #   botocore
    #   requests
    #   sentry-sdk
wandb==0.19.8
    # via -r requirements.in
wheel==0.45.1
    # via pip-tools
xxhash==3.5.0
    # via mosaicml-streaming
yasoo==0.12.6
    # via -r requirements.in
zstd==1.5.6.6
    # via mosaicml-streaming

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
